Move context into its own module
